Added store verification, and modified seller class with new method checkStores

+++ src/Seller.java
@@ -20,23 +20,41 @@ public class Seller extends Account {
         }
     }
 
-    public Seller(Account account) {
-
+    public Seller(Account account) throws IOException {
+        super(account);
+        f = new File(super.getUsername() + ".txt");
+        pw = new PrintWriter(new BufferedWriter(new FileWriter(super.getUsername() + ".txt")));
+        br = new BufferedReader(new FileReader(super.getUsername() + ".txt"));
+        String line = "";
+        while ((line = br.readLine()) != null) {
+            stores.add(line);
+        }
     }
 
-    public void createStore(String sName) throws IOException {
-        //add to list of stores
-        stores.add(sName);
-        //delete the file;
-        f.delete();
-        //remake the file;
-        pw = new PrintWriter(new BufferedWriter(new FileWriter(super.getUsername() + ".txt")));
-        //append the file with each line being the store in the list;
-        for (String s : stores) {
-            pw.println(s);
-            pw.flush();
+    public boolean createStore(String sName) throws IOException {
+        boolean storeFound = checkStores(sName);
+
+        if (!storeFound) {    // If checkStore does not find store in list
+            //add to list of stores
+            stores.add(sName);
+            //delete the file;
+            f.delete();
+            //remake the file;
+            pw = new PrintWriter(new BufferedWriter(new FileWriter(super.getUsername() + ".txt")));
+            //append the file with each line being the store in the list;
+            for (String s : stores) {
+                pw.println(s);
+                pw.flush();
+            }
+            pw.close();
+
+            PrintWriter pwStores = new PrintWriter(new FileWriter("stores.txt"));
+            pwStores.println(sName);
+            pwStores.close();
+            return true;
+        } else {
+            return false;
         }
-        pw.close();
     }
 
     public void viewCustomers() {
@@ -46,4 +64,19 @@ public class Seller extends Account {
     public boolean searchCustomer() {
         return true;
     }
+
+    public boolean checkStores(String storeName) throws IOException {
+        BufferedReader bfrStores = new BufferedReader(new FileReader(new File("stores.txt")));
+        String line = bfrStores.readLine();
+
+        while (line != null) {
+            if (line.equals(storeName)) {
+                return true;
+            }
+            line = bfrStores.readLine();
+        }
+        return false;
+    }
+
+
 }
